+++ src/backends/csharp/gen_visitor.rs
@@ -560,6 +560,7 @@ mod tests {
                 is_copy: false,
                 has_serde: true,
                 has_default: false,
+                serde_content: None,
                 serde_tag: None,
                 serde_untagged: false,
                 serde_rename_all: Some("snake_case".to_string()),