+++ src/backends/java/gen_visitor/files.rs
@@ -714,6 +714,7 @@ pub(super) mod tests {
                 is_copy: false,
                 has_serde: true,
                 has_default: false,
+                serde_content: None,
                 serde_tag: None,
                 serde_untagged: false,
                 serde_rename_all: None,